fix dockerfile for postgres, add postgres flags
