+++ cmd/secret/cmd/root_test.go
@@ -3,7 +3,6 @@ package cmd
 import (
 	"bytes"
 	"context"
-	"encoding/json"
 	"fmt"
 	"io/ioutil"
 	"net"
@@ -15,168 +14,9 @@ import (
 	"testing"
 	"time"
 
-	"github.com/go-redis/redis/v8"
-
 	api "github.com/go-itools-internship/go-secret/pkg/http"
 	"github.com/stretchr/testify/require"
 )
-
-func TestRoot_Set(t *testing.T) {
-	t.Run("expect one keys", func(t *testing.T) {
-		key := "key value"
-		path := "testFile.txt"
-		ctx, cancel := context.WithTimeout(context.Background(), 3*time.Second)
-		defer cancel()
-
-		r := New()
-		r.cmd.SetArgs([]string{"set", "--key", key, "--value", "test value", "--cipher-key", "ck", "--path", path})
-		err := r.Execute(ctx)
-		require.NoError(t, err)
-		defer func() {
-			require.NoError(t, os.Remove(path))
-		}()
-
-		testFile, err := os.Open(path)
-		require.NoError(t, err)
-		defer func() {
-			require.NoError(t, testFile.Close())
-		}()
-
-		fileData := make(map[string]string)
-		require.NoError(t, json.NewDecoder(testFile).Decode(&fileData))
-
-		var got string
-		require.Len(t, fileData, 1)
-		for key := range fileData {
-			got = key
-			break // we iterate one time to get first key
-		}
-		require.EqualValues(t, key, got)
-	})
-	t.Run("expect set data only redis storage", func(t *testing.T) {
-		key := "12345"
-		path := ""
-		redisURL := "localhost:6379"
-		ctx, cancel := context.WithTimeout(context.Background(), 3*time.Second)
-		defer cancel()
-
-		r := New()
-		r.cmd.SetArgs([]string{"set", "--key", key, "--value", "test value", "--cipher-key", "ck", "--redis-url", redisURL})
-		err := r.Execute(ctx)
-		require.NoError(t, err)
-
-		_, err = os.Open(path)
-		require.Error(t, err)
-
-		rdb := redis.NewClient(&redis.Options{Addr: redisURL, Password: "", DB: 0})
-
-		val, err := rdb.Get(ctx, key).Result()
-		require.NoError(t, err)
-		require.NotEmpty(t, val)
-	})
-
-	t.Run("expect two keys", func(t *testing.T) {
-		firstKey := "first key"
-		secondKey := "second key"
-		path := "testFile.txt"
-		ctx, cancel := context.WithTimeout(context.Background(), 3*time.Second)
-		defer cancel()
-
-		r := New()
-		r.cmd.SetArgs([]string{"set", "--key", firstKey, "--value", "test value", "--cipher-key", "ck", "--path", path})
-		err := r.Execute(ctx)
-		require.NoError(t, err)
-
-		r2 := New()
-		r2.cmd.SetArgs([]string{"set", "--key", secondKey, "--value", "test value", "--cipher-key", "ck", "--path", path})
-		err = r2.Execute(ctx)
-		require.NoError(t, err)
-
-		require.NoError(t, err)
-		defer func() {
-			require.NoError(t, os.Remove(path))
-		}()
-
-		testFile, err := os.Open(path)
-		require.NoError(t, err)
-		defer func() {
-			require.NoError(t, testFile.Close())
-		}()
-
-		fileData := make(map[string]string)
-		require.NoError(t, json.NewDecoder(testFile).Decode(&fileData))
-		require.Len(t, fileData, 2)
-		require.Contains(t, fileData, firstKey)
-		require.Contains(t, fileData, secondKey)
-	})
-}
-
-func TestRoot_Get(t *testing.T) {
-	key := "key-value"
-	value := "60OBdPOOkSOu6kn8ZuMuXtAPVrUEFkPREydDwY6+ip/LrAFaHSc="
-	path := "testFile.txt"
-	t.Run("success", func(t *testing.T) {
-		file, err := os.Create(path)
-		ctx, cancel := context.WithTimeout(context.Background(), 3*time.Second)
-		defer cancel()
-
-		require.NoError(t, err)
-		defer func() {
-			require.NoError(t, os.Remove(path))
-		}()
-
-		defer func() {
-			require.NoError(t, file.Close())
-		}()
-
-		fileTestData := make(map[string]string)
-		fileTestData[key] = value
-		require.NoError(t, json.NewEncoder(file).Encode(&fileTestData))
-
-		r := New()
-		r.cmd.SetArgs([]string{"get", "--key", key, "--cipher-key", "ck", "--path", path})
-		executeErr := r.Execute(ctx)
-		require.NoError(t, executeErr)
-
-		testFile, err := os.Open(path)
-		require.NoError(t, err)
-
-		defer func() {
-			require.NoError(t, testFile.Close())
-		}()
-
-		fileData := make(map[string]string)
-		require.NoError(t, json.NewDecoder(testFile).Decode(&fileData))
-		var got string
-		require.Len(t, fileData, 1)
-		for _, value := range fileData {
-			got = value
-			break // we iterate one time to get first value
-		}
-		require.EqualValues(t, value, got)
-	})
-	t.Run("success after get redis command", func(t *testing.T) {
-		redisURL := "localhost:6379"
-		ctx, cancel := context.WithTimeout(context.Background(), 3*time.Second)
-		defer cancel()
-
-		r := New()
-
-		r.cmd.SetArgs([]string{"set", "--key", key, "--value", "test value", "--cipher-key", "ck", "--redis-url", redisURL})
-		executeErr := r.Execute(ctx)
-		require.NoError(t, executeErr)
-
-		var b bytes.Buffer
-		r.cmd.SetOut(&b)
-
-		r.cmd.SetArgs([]string{"get", "--key", key, "--cipher-key", "ck", "--redis-url", redisURL})
-		err := r.Execute(ctx)
-		require.NoError(t, err)
-		out := b.String()
-		require.NoError(t, err)
-		require.EqualValues(t, "test value\n", out)
-	})
-}
 
 func TestRoot_Server(t *testing.T) {
 	key := "key-value"
@@ -203,6 +43,38 @@ func TestRoot_Server(t *testing.T) {
 			require.EqualValues(t, http.StatusNoContent, resp.StatusCode)
 			require.NoError(t, resp.Body.Close())
 		})
+		t.Run("expect postgres set method success", func(t *testing.T) {
+			expectedSipherKey := "key value"
+			postgresURL := "localhost:5432"
+			ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
+			defer cancel()
+
+			port, err := GetFreePort()
+			require.NoError(t, err)
+			r := New()
+			r.cmd.SetArgs([]string{"server", "--postgres-url", postgresURL})
+			go func() {
+				err := r.Execute(ctx)
+				if err != nil {
+					fmt.Println(err)
+				}
+			}()
+			time.Sleep(2 * time.Second)
+			defer func() {
+				require.NoError(t, os.Remove("file.txt"))
+			}()
+
+			client := http.Client{Timeout: time.Second}
+			body := bytes.NewBufferString(`{"getter":"key-value","method":"remote","value":"test-value-1"}`)
+			req := httptest.NewRequest(http.MethodPost, "http://localhost:"+strconv.Itoa(port), body)
+			req.Header.Set(api.ParamCipherKey, expectedSipherKey)
+			req.RequestURI = ""
+
+			resp, err := client.Do(req)
+			require.NoError(t, err)
+			require.EqualValues(t, http.StatusNoContent, resp.StatusCode)
+			require.NoError(t, resp.Body.Close())
+		})
 		t.Run("expect redis set method success", func(t *testing.T) {
 			expectedSipherKey := "key value"
 			redisURL := "localhost:6379"
@@ -220,9 +92,9 @@ func TestRoot_Server(t *testing.T) {
 				}
 			}()
 			time.Sleep(2 * time.Second)
-			defer func() {
-				require.NoError(t, os.Remove("file.txt"))
-			}()
+			//defer func() {
+			//	require.NoError(t, os.Remove("file.txt"))
+			//}()
 
 			client := http.Client{Timeout: time.Second}
 			body := bytes.NewBufferString(`{"getter":"key-value","method":"remote","value":"test-value-1"}`)
